# openapi/opencode.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "opencode",
    "description": "opencode api",
    "version": "0.0.3"
  },
  "paths": {
...
    "/auth/{providerID}": {
      "put": {
        "operationId": "auth.set",
        "summary": "Set auth credentials",
        "description": "Set authentication credentials",
        "responses": {
          "200": {
            "description": "Successfully set authentication credentials",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BadRequestError" }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "providerID",
            "schema": { "type": "string" },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Auth" }
            }
          }
        }
      },
      "delete": {
        "operationId": "auth.remove",
        "summary": "Remove auth credentials",
        "description": "Remove authentication credentials",
        "responses": {
          "200": {
            "description": "Successfully removed authentication credentials",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BadRequestError" }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "providerID",
            "schema": { "type": "string" },
            "required": true
          }
        ]
      }
    },
...
    "/config/providers": {
      "get": {
        "operationId": "config.providers",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "List config providers",
        "description": "Get a list of all configured AI providers and their default models.",
        "responses": {
          "200": {
            "description": "List of providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "providers": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Provider" }
                    },
                    "default": {
                      "type": "object",
                      "propertyNames": { "type": "string" },
                      "additionalProperties": { "type": "string" }
                    }
                  },
                  "required": ["providers", "default"]
                }
              }
            }
          }
        }
      }
    },
...
    "/provider": {
      "get": {
        "operationId": "provider.list",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "List providers",
        "description": "Get a list of all available AI providers, including both available and connected ones.",
        "responses": {
          "200": {
            "description": "List of providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "all": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "api": { "type": "string" },
                          "name": { "type": "string" },
                          "env": {
                            "type": "array",
                            "items": { "type": "string" }
                          },
                          "id": { "type": "string" },
                          "npm": { "type": "string" },
                          "models": {
                            "type": "object",
                            "propertyNames": { "type": "string" },
                            "additionalProperties": {
                              "type": "object",
                              "properties": {
                                "id": { "type": "string" },
                                "name": { "type": "string" },
                                "family": { "type": "string" },
                                "release_date": { "type": "string" },
                                "attachment": { "type": "boolean" },
                                "reasoning": { "type": "boolean" },
                                "temperature": { "type": "boolean" },
                                "tool_call": { "type": "boolean" },
                                "interleaved": {
                                  "anyOf": [
                                    { "type": "boolean", "const": true },
                                    {
                                      "type": "object",
                                      "properties": {
                                        "field": {
                                          "type": "string",
                                          "enum": [
                                            "reasoning_content",
                                            "reasoning_details"
                                          ]
                                        }
                                      },
                                      "required": ["field"],
                                      "additionalProperties": false
                                    }
                                  ]
                                },
                                "cost": {
                                  "type": "object",
                                  "properties": {
                                    "input": { "type": "number" },
                                    "output": { "type": "number" },
                                    "cache_read": { "type": "number" },
                                    "cache_write": { "type": "number" },
                                    "context_over_200k": {
                                      "type": "object",
                                      "properties": {
                                        "input": { "type": "number" },
                                        "output": { "type": "number" },
                                        "cache_read": { "type": "number" },
                                        "cache_write": { "type": "number" }
                                      },
                                      "required": ["input", "output"]
                                    }
                                  },
                                  "required": ["input", "output"]
                                },
                                "limit": {
                                  "type": "object",
                                  "properties": {
                                    "context": { "type": "number" },
                                    "input": { "type": "number" },
                                    "output": { "type": "number" }
                                  },
                                  "required": ["context", "output"]
                                },
                                "modalities": {
                                  "type": "object",
                                  "properties": {
                                    "input": {
                                      "type": "array",
                                      "items": {
                                        "type": "string",
                                        "enum": [
                                          "text",
                                          "audio",
                                          "image",
                                          "video",
                                          "pdf"
                                        ]
                                      }
                                    },
                                    "output": {
                                      "type": "array",
                                      "items": {
                                        "type": "string",
                                        "enum": [
                                          "text",
                                          "audio",
                                          "image",
                                          "video",
                                          "pdf"
                                        ]
                                      }
                                    }
                                  },
                                  "required": ["input", "output"]
                                },
                                "experimental": { "type": "boolean" },
                                "status": {
                                  "type": "string",
                                  "enum": ["alpha", "beta", "deprecated"]
                                },
                                "options": {
                                  "type": "object",
                                  "propertyNames": { "type": "string" },
                                  "additionalProperties": {}
                                },
                                "headers": {
                                  "type": "object",
                                  "propertyNames": { "type": "string" },
                                  "additionalProperties": { "type": "string" }
                                },
                                "provider": {
                                  "type": "object",
                                  "properties": {
                                    "npm": { "type": "string" },
                                    "api": { "type": "string" }
                                  }
                                },
                                "variants": {
                                  "type": "object",
                                  "propertyNames": { "type": "string" },
                                  "additionalProperties": {
                                    "type": "object",
                                    "propertyNames": { "type": "string" },
                                    "additionalProperties": {}
                                  }
                                }
                              },
                              "required": [
                                "id",
                                "name",
                                "release_date",
                                "attachment",
                                "reasoning",
                                "temperature",
                                "tool_call",
                                "limit",
                                "options"
                              ]
                            }
                          }
                        },
                        "required": ["name", "env", "id", "models"]
                      }
                    },
                    "default": {
                      "type": "object",
                      "propertyNames": { "type": "string" },
                      "additionalProperties": { "type": "string" }
                    },
                    "connected": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  },
                  "required": ["all", "default", "connected"]
                }
              }
            }
          }
        }
      }
    },
    "/provider/auth": {
      "get": {
        "operationId": "provider.auth",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "Get provider auth methods",
        "description": "Retrieve available authentication methods for all AI providers.",
        "responses": {
          "200": {
            "description": "Provider auth methods",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "propertyNames": { "type": "string" },
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ProviderAuthMethod"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/provider/{providerID}/oauth/authorize": {
      "post": {
        "operationId": "provider.oauth.authorize",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "workspace",
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "providerID",
            "schema": { "type": "string" },
            "required": true,
            "description": "Provider ID"
          }
        ],
        "summary": "OAuth authorize",
        "description": "Initiate OAuth authorization for a specific AI provider to get an authorization URL.",
        "responses": {
          "200": {
            "description": "Authorization URL and method",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderAuthAuthorization"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BadRequestError" }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "method": {
                    "description": "Auth method index",
                    "type": "number"
                  },
                  "inputs": {
                    "description": "Prompt inputs",
                    "type": "object",
                    "propertyNames": { "type": "string" },
                    "additionalProperties": { "type": "string" }
                  }
                },
                "required": ["method"]
              }
            }
          }
        }
      }
    },
...
    "/tui/control/response": {
      "post": {
        "operationId": "tui.control.response",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": { "type": "string" }
          },
          { "in": "query", "name": "workspace", "schema": { "type": "string" } }
        ],
        "summary": "Submit TUI response",
        "description": "Submit a response to the TUI request queue to complete a pending request.",
        "responses": {
          "200": {
            "description": "Response submitted successfully",
            "content": {
              "application/json": { "schema": { "type": "boolean" } }
            }
          }
        },
        "requestBody": { "content": { "application/json": { "schema": {} } } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BadRequestError": {
        "type": "object",
        "properties": {
          "data": {},
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "propertyNames": { "type": "string" },
              "additionalProperties": {}
            }
          },
          "success": { "type": "boolean", "const": false }
        },
        "required": ["data", "errors", "success"]
      },
      "OAuth": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "const": "oauth" },
          "refresh": { "type": "string" },
          "access": { "type": "string" },
          "expires": { "type": "number" },
          "accountId": { "type": "string" },
          "enterpriseUrl": { "type": "string" }
        },
        "required": ["type", "refresh", "access", "expires"]
      },
      "ApiAuth": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "const": "api" },
          "key": { "type": "string" }
        },
        "required": ["type", "key"]
      },
      "WellKnownAuth": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "const": "wellknown" },
          "key": { "type": "string" },
          "token": { "type": "string" }
        },
        "required": ["type", "key", "token"]
      },
      "Auth": {
        "anyOf": [
          { "$ref": "#/components/schemas/OAuth" },
          { "$ref": "#/components/schemas/ApiAuth" },
          { "$ref": "#/components/schemas/WellKnownAuth" }
        ]
      },
...
      "Provider": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "source": {
            "type": "string",
            "enum": ["env", "config", "custom", "api"]
          },
          "env": { "type": "array", "items": { "type": "string" } },
          "key": { "type": "string" },
          "options": {
            "type": "object",
            "propertyNames": { "type": "string" },
            "additionalProperties": {}
          },
          "models": {
            "type": "object",
            "propertyNames": { "type": "string" },
            "additionalProperties": { "$ref": "#/components/schemas/Model" }
          }
        },
        "required": ["id", "name", "source", "env", "options", "models"]
      },
...
      "ProviderAuthAuthorization": {
        "type": "object",
        "properties": {
          "url": { "type": "string" },
          "method": {
            "anyOf": [
              { "type": "string", "const": "auto" },
              { "type": "string", "const": "code" }
            ]
          },
          "instructions": { "type": "string" }
        },
        "required": ["url", "method", "instructions"]
      },
...
    }
  }
}